Update Program.cs
1)	Is the code properly documented/commented?
The code has no commenting at all and may be hard to understand if a coder continued in this way. If there are any comments, I have not seen them.

2)	Does the code handle errors properly?
The code runs really well, including the WrongConcertException, specifically designed to catch an incorrect Concert. Some small issues may still slip by since this is the only custom Exception I have found, for example I'm not sure if a negatively priced seat would be caught.

3)	What suggestions do you have to improve the code?
Making comments on the code to make it easier to understand at a glance would be appreciated. Adding error handling for some cases such as negative ticket price, things like two people having the same seat, maybe even error handing different seating arrangements (for example, one concert may number each seat individually, so seats could be like "1", "23", "257" and so on, whereas other concerts could note rows and columns separately, such as "A1", "B2", C4", "AR17" and so on, or maybe even "1A", "5R" and so on, if the seats are labelled inversely. This error handing would make sure the right seat type was given if one is specified) would make it more sustainable and better if it was continued and implemented in a professional manner.